refactor: unify preview accordion
